+++ pkg/ctrlserver/graph/model/models_gen.go
@@ -7,9 +7,8 @@ import (
 	"io"
 	"strconv"
 
-	v1 "k8s.io/api/core/v1"
-
 	"github.com/chaos-mesh/chaos-mesh/api/v1alpha1"
+	"k8s.io/api/core/v1"
 )
 
 type Fd struct {
