+++ pkg/webhook/affected_namespaces.go
@@ -20,8 +20,8 @@ import (
 
 	"k8s.io/apimachinery/pkg/util/validation/field"
 
+	"github.com/chaos-mesh/chaos-mesh/api/genericwebhook"
 	"github.com/chaos-mesh/chaos-mesh/api/v1alpha1"
-	"github.com/chaos-mesh/chaos-mesh/api/v1alpha1/genericwebhook"
 )
 
 func affectedNamespaces(obj interface{}) (bool, map[string]struct{}) {
